# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Anime App API",
    "version": "2.0.0",
    "description": "API REST para personajes de anime y categorias personalizadas."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Personaje": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "poder_tecnica": {
            "type": "string"
          },
          "anime": {
            "type": "string"
          },
          "imagenes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "AnimeCategoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "genero": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "estado": {
            "type": "string",
            "enum": [
              "pendiente",
              "viendo",
              "completado"
            ]
          },
          "calificacion": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/api/{anime}/{nombre}": {
      "get": {
        "summary": "Buscar personaje por nombre",
        "tags": [
          "Personajes"
        ],
        "parameters": [
          {
            "name": "anime",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "saint-seiya",
                "hunter-x-hunter",
                "one-piece"
              ]
            }
          },
          {
            "name": "nombre",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personaje encontrado"
          },
          "404": {
            "description": "No encontrado"
          }
        }
      }
    },
    "/api/{anime}": {
      "get": {
        "summary": "Todos los personajes de un anime",
        "tags": [
          "Personajes"
        ],
        "parameters": [
          {
            "name": "anime",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "saint-seiya",
                "hunter-x-hunter",
                "one-piece"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de personajes"
          }
        }
      }
    },
    "/api/categorias": {
      "get": {
        "summary": "Listar mis categorias",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de categorias"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "post": {
        "summary": "Crear nueva categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoria creada"
          }
        }
      }
    },
    "/api/categorias/{id}": {
      "put": {
        "summary": "Actualizar categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoria actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminada"
          }
        }
      }
    },
    "/api/categorias/{id}/animes": {
      "get": {
        "summary": "Listar animes de una categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de animes"
          }
        }
      },
      "post": {
        "summary": "Agregar anime a una categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimeCategoria"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Anime agregado"
          }
        }
      }
    },
    "/api/categorias/{id}/animes/{animeId}": {
      "put": {
        "summary": "Actualizar anime de una categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime actualizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar anime de una categoria",
        "tags": [
          "Categorias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado"
          }
        }
      }
    }
  },
  "tags": []
}